detect: allow for more than 64k mpm rules

+++ src/util-mpm-ac-bs.h
@@ -88,7 +88,7 @@ typedef struct SCACBSCtx_ {
 
     /* the size of each state */
     uint16_t single_state_size;
-    uint16_t max_pat_id;
+    uint32_t max_pat_id;
 } SCACBSCtx;
 
 typedef struct SCACBSThreadCtx_ {
